Neaten facade--move project system class into subfolder, sweeten identifiers.
