+++ src/NuGet.Clients/VisualStudio.Facade/NuGet.VisualStudio.Facade.csproj
@@ -68,9 +68,9 @@
     <Reference Include="System.ComponentModel.Composition" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Telemetry\ITelemetrySession.cs" />
-    <Compile Include="ProjectHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ProjectSystem\ProjectHelper.cs" />
+    <Compile Include="Telemetry\ITelemetrySession.cs" />
     <Compile Include="Telemetry\TelemetryEvent.cs" />
     <Compile Include="Telemetry\TelemetrySession.cs" />
   </ItemGroup>
